unify: introduce `PropTycker` for `Prop`-related
