+++ base/src/main/java/org/aya/tyck/tycker/UnifiedTycker.java
@@ -43,7 +43,7 @@ import java.util.function.Function;
  * @see #checkBoundaries
  * @see #inheritFallbackUnify
  */
-public sealed abstract class UnifiedTycker extends MockedTycker permits ExprTycker {
+public sealed abstract class UnifiedTycker extends MockedTycker permits PropTycker {
   protected UnifiedTycker(@NotNull Reporter reporter, Trace.@Nullable Builder traceBuilder, @NotNull TyckState state) {
     super(reporter, traceBuilder, state);
   }
